Stub "RDRAM size" value in RDRAM to fix bug described in previous commit message

+++ src/n64/core/Mem.hpp
@@ -17,21 +17,21 @@ struct Mem {
   }
 
   template <bool tlb = true>
-  u8 Read8(Registers&, u32, s64);
+  u8 Read8(Registers&, u64, s64);
   template <bool tlb = true>
-  u16 Read16(Registers&, u32, s64);
+  u16 Read16(Registers&, u64, s64);
   template <bool tlb = true>
-  u32 Read32(Registers&, u32, s64);
+  u32 Read32(Registers&, u64, s64);
   template <bool tlb = true>
-  u64 Read64(Registers&, u32, s64);
+  u64 Read64(Registers&, u64, s64);
   template <bool tlb = true>
-  void Write8(Registers&, u32, u32, s64);
+  void Write8(Registers&, u64, u32, s64);
   template <bool tlb = true>
-  void Write16(Registers&, u32, u32, s64);
+  void Write16(Registers&, u64, u32, s64);
   template <bool tlb = true>
-  void Write32(Registers&, u32, u32, s64);
+  void Write32(Registers&, u64, u32, s64);
   template <bool tlb = true>
-  void Write64(Registers&, u32, u64, s64);
+  void Write64(Registers&, u64, u64, s64);
 
   MMIO mmio;
   u8 pifRam[PIF_RAM_SIZE]{};
@@ -49,5 +49,5 @@ private:
 };
 
 template <bool tlb = true>
-bool MapVAddr(Registers& regs, TLBAccessType accessType, u32 vaddr, u32& paddr);
+bool MapVAddr(Registers& regs, TLBAccessType accessType, u64 vaddr, u32& paddr);
 }
